added save file and fixed load function
