+++ Milestone4/Memory.py
@@ -16,16 +16,11 @@ class Memory:
 
     def readProgram(self, file):
         """Reads in file containing a basicML program. Loads registers."""
-        with open(file, "r") as program:
-            lst = program.readlines()
-            for i in range(len(lst)):
-                self.registers[i] = lst[i].strip()
-        return
-
-    def readTextArea(self, input):
-        lst = input.split('/n')
+        lst = file.split("\n")
         for i in range(len(lst)):
             self.registers[i] = lst[i].strip()
+        return
+
     # i/o operations:
     # READ
     def read(self, targetMemoryLocation, word):
